Frontend code reorg
